refactor: mv root into routers dir
