+++ packages/api/src/index.ts
@@ -1,7 +1,7 @@
 import type { inferRouterInputs, inferRouterOutputs } from "@trpc/server";
 
-import type { AppRouter } from "./root";
-import { appRouter } from "./root";
+import type { AppRouter } from "./router/root";
+import { appRouter } from "./router/root";
 import { createCallerFactory, createTRPCContext } from "./trpc";
 
 /**
